seprate validation from output
